Added Windbg's TTD.exe (1.11.443.0) bound together with its dependencies using Enigma Virtual Box.
This uses the Enigma Virtual Box executable binder over at
https://enigmaprotector.com/assets/files/enigmavb.exe to bind Windbg's
`TTD.exe` with all of its dependencies so that it can be deployed as a
single file. The filename of each variation correlates to the required
architecture. This results in `ttd.amd64.exe` for 64-bit, `ttd.x86.exe`
for 32-bit, and `ttd.wow64.exe` for emulated 32-bit.

Each executable contains the entirety of the matching TTD directory.
This includes the "Analyzers" sub-directory which contains the
javascript that is used by the TTD client.

In regards to the configuration for these bound executables, each
executable was configured so that its executables and DLL dependencies
are always written to disk. This satisfies the requirements of the TTD
client which needs its files on disk in order to inject them into the
target process.